Testing Products and Categories

+++ Src/DTO/Product/EditProductRequest.cs
@@ -11,7 +11,7 @@ namespace Aduaba.DTO.Product
         public string NewName { get; set; }
         public string NewDescription { get; set; }
         
-        public double NewPrice { get; set; }
+        public decimal NewPrice { get; set; }
         
         public string NewImageUrl { get; set; }
         public int UpdatedQuantity { get; set; }
